# Pull Request Checklist

* [x] Have you read through the [contributor guidelines](https://www.playframework.com/contributing)?
* [x] Have you signed the [Typesafe CLA](https://www.typesafe.com/contribute/cla)?
* [x] Have you [squashed your commits](https://www.playframework.com/documentation/latest/WorkingWithGit#Squashing-commits)?
* [x] Have you added copyright headers to new files?
* [x] Have you checked that both Scala and Java APIs are updated?
* [x] Have you updated the documentation for both Scala and Java sections?
* [x] Have you added tests for any changed functionality?

## Purpose

Optimizes `StandaloneAhcWSRequest.add*` methods from O(n) down to O(eC) by avoiding reconstructing maps from scratch.

## Background Context
I spotted this while benching https://github.com/playframework/play-ws/pull/288.

`StandaloneAhcWSRequest` uses immutable maps that support structural sharing, yet the implementations of `addQueryStringParameters((String, String)*)` and `addHttpHeaders(hdrs: (String, String)*)` rebuild the entire map on each call. For a request with *n* elements, appending a new one is an O(*n*) operation over an O(eC) data structure. Inserting *n* elements is O(*n²*).

Play 2.5 `WSRequest` has an O(eC) implementation, so this seems like new behavior in the pre-release of play-ahc-ws-standalone v2.

## Benchmarks
`AddQueryParamsBench` measures adding a single query param to a request that already has *size* query params.

Before
```
[info] Benchmark                      (size)  Mode  Cnt        Score       Error  Units
[info] AddQueryParamsBench.addParams       1  avgt   10      417.620 ±    17.237  ns/op
[info] AddQueryParamsBench.addParams      10  avgt   10     2257.053 ±    76.440  ns/op
[info] AddQueryParamsBench.addParams     100  avgt   10    18839.600 ±   801.003  ns/op
[info] AddQueryParamsBench.addParams    1000  avgt   10   224491.116 ± 12133.870  ns/op
[info] AddQueryParamsBench.addParams   10000  avgt   10  3065837.189 ± 82797.559  ns/op
```

We cross the 1s/call threshold somewhere between the 1,000 and 10,000th param. At 10k query params, constructing the whole request takes 16 seconds.

Admittedly 10k query params is on the high end of what we're likely to see in real apps, but we can also bring it down to nanoseconds pretty trivially.

After
```
[info] Benchmark                      (size)  Mode  Cnt    Score   Error  Units
[info] AddQueryParamsBench.addParams       1  avgt   10   47.978 ± 0.388  ns/op
[info] AddQueryParamsBench.addParams      10  avgt   10   76.495 ± 0.663  ns/op
[info] AddQueryParamsBench.addParams     100  avgt   10  118.205 ± 1.072  ns/op
[info] AddQueryParamsBench.addParams    1000  avgt   10  141.506 ± 1.681  ns/op
[info] AddQueryParamsBench.addParams   10000  avgt   10  151.452 ± 2.560  ns/op
```

## Caveats
There is a default implementation of these methods in the `trait StandaloneWSRequest`. I don't see any harm in leaving them there.

## References

No other issues or PRs that I'm aware of.