@marcospereira 2019-02-14
Backport to 2.0.x: 5bd1434f36ede6c00cefbe1b534017b75e39c89b